% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncPlot.r
\name{AncPlot}
\alias{AncPlot}
\title{Function for mapping ie results onto a phylogenetic tree}
\usage{
AncPlot(AncObject, tree, branchWidth, branchCutoff, labelSize, scaling)
}
\arguments{
\item{AncObject}{object comprising the output of an ie analysis}
}
\value{
phylogenetic tree with mapped change and ancestral states
}
\description{
This function maps the results of an ie analysis onto a tree
}

