% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateDistr.r
\name{RateDistr}
\alias{RateDistr}
\title{Function to allocate the rate distribution between 2 variables to rate quadrants}
\usage{
RateDistr(Y, X, neutral)
}
\arguments{
\item{Y}{AncObject of variable 1}

\item{X}{AncObject of variable 2}
}
\value{
list of ratedistribution elements and branches for each quadrant
}
\description{
This function allocates the rate distribution between 2 variables to rate quadrants
}

